Expose `cmap` in `plot_crosstab()` (#3)
* Expose `cmap` in crosstab plots.

Include `cmap` as a parameter of `plot_crosstab` which is then passed to
`sns.color_palette`. This colour map is used to create the colour bar,
too.

* Update demo notebook.

Rerun the notebook with a crosstab plot that uses a custom colour map.
In this case, it's one of the custom `seaborn` ones: `light:coral`.

This does cause a couple of `FutureWarning`s. Likely because I've
recently updated my `pandas`. Will fix in this branch as it's quick.

* Future-proof plotting functions.

Dealt with those `FutureWarning` instances. There may be more in the
codebase.

Co-authored-by: Henry Wilde <henry.wilde@ons.gov.uk>