Property-based tests for the `Evaluator` class (#12)
* Start writing tests for Evaluator class.

* Fix load and save tests.

* Finish Evaluator tests (ex. plotting)

Hypothesis tests for plots are arduous and memory-heavy. These functions
act as wrappers to the actual plotting functions so I don't see an
immediate need to test them directly.

* Expand Evaluator test for results.

* Add hypothesis to dev requirements.

* Unify metric interface.

Previously, metric names would only be recorded if an alias was
provided. This led to a few oddly complicated logic blocks to resolve.

I have made it so metric name is always recorded and becomes the default
alias if none is provided.

* Remove old Evaluator tests

* Add config for coverage and hypothesis aux files

Ignore their generated aux files and append coverage settings to
`pyproject.toml`