Fix bug in `feature_density_diff`
Previously, if `real` and `synth` didn't have identical value sets for
`feature` then their encoded selves would not be equivalent, making the
calculated density difference ineffective.

A follow-on from this change is the change in the evaluator tests
(kruskal need distinct datasets to run). Is there a catch to put in
there?