Property-based tests for utility functions (#13)
* Write tests and fix logic for combine/separate fns

Two property-based tests for the `df_combine` and `df_separate`
functions. To write these tests, I have had to slightly alter the logic
of the separate function; previously, you couldn't easily achieve the
full functionality between the parameters.

Also, acting on a single "feature" list was causing issues with the testing suite.

* Write `launder` test

* Write tests for `cat_encode`.

This function is quite complex so I've written three tests to cover some
core scenarios while covering the parameters. Ideally, this could be
refactored so testing can be modularised fully.

* Rework column logic in `cat_encode`.

There was a logical hole here where you couldn't pass and receive only
numeric columns even with `return_all=True`. This small change makes it so
the `return_all` parameter determines whether we return every column or
just those we want encoded.

* Add `cat_encode` test for `convert_only`

Sometimes you might not want a dataset to be integer-encoded--that's why
the `convert_only` parameter exists.

The added test checks the functionality of that parameter but does not
add to the coverage.

* Write test for `feature_density_diff`

Write a (failing) test for this function that behaves as we would expect
it to.

* Fix bug in `feature_density_diff`

Previously, if `real` and `synth` didn't have identical value sets for
`feature` then their encoded selves would not be equivalent, making the
calculated density difference ineffective.

A follow-on from this change is the change in the evaluator tests
(kruskal need distinct datasets to run). Is there a catch to put in
there?