Add test and catch for categoric Q-Q plots.
We shouldn't be allowed to make Q-Q plots for categories as you can't
linearly interpolate between categories. This stops that from happening.

As an aside, there is no need for the `suggest_label_rotation()`
function now since large numeric values are written in scientific
notation. It also can't be tested for this reason.