Refactor small parts of `correlation.py`
The main changes are:

- `sklearn.metrics.mean_squared_error()` only works on arrays with
  specific shapes. Calculating MSE for numerics and arrays is
  straightforward enough.
- The correlation ratio function would fall over if NaNs were included
  in either series. These are now dropped first, and the docstring has
  been updated.

I'm not sure how I feel about the repeated checks on `feats` throughout.
They aren't equivalent, and the policy on throwing errors or sending
warnings is not consistent. For example, no warning is issued for
`correlation_MSD()` despite only being valid for numeric columns.
Meanwhile, one is sent for `cramers_v_MSE()` when numeric columns are
included.

My preference would be to avoid the warnings and update the
documentation to specify that integer-encoded variables should be cast
as objects or categories first.

Also, there are issues with consistency between mean-squared error and
mean-squared difference throughout.