Property-based tests for the `metrics` subpackage (#16)
* Move `pytest` fixtures to `conftest.py`

This is the preferred way to make fixtures available across files.

* Move old metric tests to "example" tests

* Write property-based tests for classifier metrics

* Modularise classification metric wrapper

The old wrapper was a little hard to read (and to test) so I have broken
it down into parts where it seems straight forward to. The tests already
reflect these changes.

* Remove dataset fixtures from `utils.py`

* Write property-based tests for `metrics/cluster.py`

* Modularise `cluster.py`

Again, the functions were ripe for breaking down into smaller, readable,
testable pieces.

* Write tests for `correlation.py`

* Refactor small parts of `correlation.py`

The main changes are:

- `sklearn.metrics.mean_squared_error()` only works on arrays with
  specific shapes. Calculating MSE for numerics and arrays is
  straightforward enough.
- The correlation ratio function would fall over if NaNs were included
  in either series. These are now dropped first, and the docstring has
  been updated.

I'm not sure how I feel about the repeated checks on `feats` throughout.
They aren't equivalent, and the policy on throwing errors or sending
warnings is not consistent. For example, no warning is issued for
`correlation_MSD()` despite only being valid for numeric columns.
Meanwhile, one is sent for `cramers_v_MSE()` when numeric columns are
included.

My preference would be to avoid the warnings and update the
documentation to specify that integer-encoded variables should be cast
as objects or categories first.

Also, there are issues with consistency between mean-squared error and
mean-squared difference throughout.

* Fix num columns bug in `datasets()` strategy

* Write property-based tests for univariate metrics.

* Clean up feature logic in univariate metrics.

* Generalise `isinstance()` calls for stats tests

Some of the hypothesis tests used as univariate metrics have been
refactored in recent versions of the `scipy` library. To ensure
test-comparability between versions, we generalise to check the `repr`
of the result rather than its class.

Also, clean up the messy logic in the Kullback-Leibler test.

* Remove upper bound from KL test

Kullback-Leibler is not bounded above by one (as suggested by the
documentation). I've now removed this bound and the test passes
consistently.

* Update `kullback_leibler()` documentation

* Add bounds to KL test

Oversized floats are causing occasional issues. Just cap them off as
they are outrageously large.

* Expand data assumption for KW test

Kruskal-Wallis requires the two datasets to be non-identical, including
duplicate rows.

* Add integer catch to tests for pval=1

* Catch weird Python 3.7 things in univariate tests

- Integers from tests were coming back as `np.int32` so did not match
  `int`
- `scipy.stats._morestats` does not exist in version for 3.7

* Write property-based tests for privacy metrics

* Fix some patchy logic in privacy metrics

- `sample_overlap_score()` allowed some cases to fall through if
  `len(real) != len(synth)`
- resolving features (again)

* Write majority of tests for `metrics/propensity.py`

Just need the wrapper test for the larger `propensity_metrics()`
function.

Also moved some reused code to `tests/utils.py`.

* Fix bugs in `propensity.py`

- feature resolution
- one of listed methods was not consistent with other functions
  (Logistic Regression)
- formulae for expected and std dev of pMSE used k rather than k-1; this
  means that they did not satisfy the required bounds for pMSE-ratio
  when using LogisticRegression (where theoretical values come from)...

* Update example values for pMSE w. LogReg

Fixing the `(k-1)` bug meant that the expected values using those
functions were incorrect. To keep those tests in the suite (is that a
good idea?), I've updated them with the values I get on my machine with
the development environment.

* Write test for propensity wrapper

* Fix feature resolution in propensity wrapper