Refactor propensity metrics
The old implementation had a number of issues including documentation
and insufficient modularisation. It also made it hard to find the source
of the randomness leak.

The overall functionality of the core `propensity_metrics()` function is
unchanged except only lists of features are allowed in `feats`. This
reflects the existing documentation.