Sort out pseudo-random number generation (#18)
* Update PRNG strategy for datasets.

* Update metric example tests.

I'm not sure I like these tests now that the PRNG strategy has changed.

Note that the propensity metric tests are still incorrect--there is some
sort of randomness leak when `method="LogisticRegression"` and that is
yet to be committed.

* Rewrite propensity metric tests

The randomness leak was hard to pinpoint, so I'm doing some refactoring.

* Refactor propensity metrics

The old implementation had a number of issues including documentation
and insufficient modularisation. It also made it hard to find the source
of the randomness leak.

The overall functionality of the core `propensity_metrics()` function is
unchanged except only lists of features are allowed in `feats`. This
reflects the existing documentation.

* Update metric example tests

* Remove lower bound from crosstab plot test

Hypothesis, for some reason, has now found this counterexample to my
poorly thought-out reasoning for the counts in these plots.

* Move from `math.comb()`

The `math.comb()` function is new in Python 3.8, so not usable for
production. Instead, using `scipy.special.binom()`, which is available
for all `scipy` versions that meet the package requirements.

* Suppress `data_too_large` check in evaluator

Recurrent failing test locally about data that is too large in the
`Evaluator.plot_histograms()` method.