Property-based tests for dataset creation (#19)
* Write tests for dataset creation

* Minor refactor of dataset creation

* Fix typo in data-label strategy

For a reason unbeknownst to me, the labels are multiplied by three when
creating data proper. This makes sure the strategy does, too.

* Scrap label tripling and update metric examples

There is no immediate reason to multiply the integer labels by three
when making a toy dataset. As such, I've removed this step (and from the
test, undoing the last commit) and updated any metric examples that
changed.

* Update example notebook