Sort out source code names (#22)
* Rename `evaluator.py`

* Rename elements of `datasets.py`

* Rename + refactor `plot.py`

Function methods are not good practice as they cannot be tested.

* Write test for pairwise Cramer's function

* Clean up some loose formatting

* Rename elements of `metrics/classification.py`

* Rename `metrics/cluster.py` items + drop MSD

The single-k clustered MSD function was already so modular that it
didn't warrant being kept separate.

* Rework correlation metrics

In renaming the correlation functions, I identified an easy opportunity
to streamline the interface. Now, all correlation MSD metrics are
contained within `metrics.correlation.correlation_msd` and the method
can be chosen with `method`. This also adds Spearman and Kendall
correlation for free (thanks, `pandas`!)

* Remove pairwise Cramer's functions from `plot.py`

* Fix upper limits for correlation MSD tests

* Rename elements of `privacy.py`

Also refactor sample generation and write a test for it.

* Move feature density difference to its own module

* Rename `univariate_distance.py` as `univariate.py`

Now that all the metrics in there *are* univariate, we're away. But they
certainly are not all distances!

* Sweep for remaining feature diff occurrences

* Update metric example tests