Add SPECKS metric
Synthetic data generation; Propensity score matching; Empirical
Comparison via the Kolmogorov-Smirnov distance.

This metric does what it says on the tin - providing us with another
propensity-based approach to measuring generic, multivariate utility.
The key benefit of this metric over the other pMSE metrics is that it
does not rely on having a large dataset to be reasonably reliable.

Details in Bowen et al (2018), Section 2.4. The paper has now been
published in Metron under
[DOI 10.1007/s40300-021-00201-0](https://doi.org/10.1007/s40300-021-00201-0).